Wwodrich/clock controls

+++ src/core/entity/building.cpp
@@ -7,7 +7,7 @@ Building::Building(Rect const& border, int id, tcod::ColorRGB color)
 
 Building::Building(const ant_proto::Building& msg)
     : border(msg.rect()),
-      id(msg.id()),
+    id(static_cast<long>(msg.id())),
       color(tcod::ColorRGB{static_cast<uint8_t>(msg.col().r()),
                            static_cast<uint8_t>(msg.col().g()),
                            static_cast<uint8_t>(msg.col().b())}) {
